# ssa: removes map use for Value aliasing